Move use-resource-simple into use-resource
